Fixed bug for `Boilerplate`
To properly map a `text` column to a Java Enum, the enum used has to be
name-only. This way, calling `toString` on it will return the name,
which corresponds to the enum itself. Previously, while we were able to
map the enum to a String value, we were unable to map back to the enum
itself. This has now been fixed.