Introducing `FallbackParser`
Extracted exception-handling behaviour from JavaParser to produce new dedicated
parser: FallbackParser. This parser can only extract the bare minimum in terms
of information, and does not rely on code structure to do so. It is used either
in the event of a ParsingException being thrown by a dedicated parser, or if
there is no parser dedicated to a specific Language. While not designed to fail
it still may produce IOExceptions in rare cases.