Added custom `Git` utilities
Due to JGit's lack of support for shallow cloning, I have opted to create my own Git
API. It currently only supports features which we have used in the crawling of code,
namely cloning and last commit info retrieval. While we are removing the JGit as a
module dependency, we may reintroduce it later for analyzing project histories for the
bug fixing dataset.