Removing the wait-service-exit script
Since the condition option works in depends_on, we can use
service_completed_successfully to wait for dl4se-liquibase to finish
before starting the dl4se-crawler service. The script was originally
created because I believed this functionality was unavailable to me.
From now on, check the correctness of the compose file using:
docker-compose config