Added new `processing` model classes
They will be used to store information about the processing part of user
-submitted tasks. The `Processing` class was written as generic as
possible, storing only an id and a reference to the task itself.
The rationale behind this was that we are still not sure about
processing operations for the other tasks. The only concrete class is
`CodeProcessing`, as abstraction and masking can be applied at both the
`File` and `Function` granularity levels.