Added new `task` model classes
This is the main mechanism behind persisting user-submitted tasks.
Each `Task` is associated to the user that submitted it, and contains
references to the `Query` and `Processing` specified by the user.
The relationship between user and task is one-to-many, while the
relationship between task and query/processing is one-to-one.
Each task is also assigned a UUID, which in comparison to the internal
database ID is actually exposed to the user, and should aid in
identifying individual tasks faster. The task also stores a `State`,
which denotes its execution phase, as well as a reference to the file
we are exporting to, an entity id checkpoint for restarts in case of
failure, and general numeric and timestamp information of progress.