Changed `hashCode` implementations for `Query` and `Processing`
Both now include the runtime type in the calculation of the hash. This
should be enough to prevent any collisions for queries of different
types (granularity levels) with similar parameters. For instance a user
can submit a file and function query with the same parameters, but no
more than one of each.