Inheritance strategy of `Query` from TABLE_PER_CLASS to SINGLE_TABLE
There was a 95% column overlap between `file_query` and `function_query`
so it made sense that they should be part of the same table. Especially
since using a single discriminator column is far better than having two
nearly identical tables that differ in only one column. The only side
effect of this is that we now have to make their exclusive columns
nullable. Given that we perform database migrations, removing the
nullability constraints of specific `query` columns, while not pretty,
will not pose any challenge.