Added `CodeRepository` method for counting the query results
This will be run prior to the task being executed. We do this to
"estimate" the total number of results that will be processed. I say
"estimate" because we do not execute an actual count query. Instead, we
get the row number of the last result set element from a scrollable
result set of the regular query.