Improved conversion from `Task` to jOOQ `Query`
Instead of converting from a `CodeTask` to a single representative jOOQ
query, we now have a converter that accepts any type of `Task`, and
converts it to the corresponding `Queries`:
- The first being a count query, used for calculating the total number
  of rows that match the user-submitted specifications
- The second being the select query, used for streaming over the
  database, matching any rows that correspond to aforementioned
  specifications

This change was necessary, because the number of results returned by the
old count query varied for the same queries. Now we can at the very
least obtain a consistent estimate, without any incurred performance
penalties.