Generified `CodeQuery` converters
Instead of using two separate converters and deciding which one to use
based on the runtime `CodeQueryDto` type, we instead simply use a
generic converter, that appropriately converts the `FileQueryDto` or
`FunctionQueryDto` to the appropriate `CodeQuery` subclass. This
improves the extensibility of the hierarchy, as we now only have to
change the converter when introducing a new `CodeQuery` subclass, as
opposed to modifying the already convoluted controller implementation.