Added scheduled task that cleans week-old files from completed tasks
At each hour mark of the day, a scheduled job is triggered that checks
for finished tasks that are older than a week, and deletes files
associated to these tasks. To prevent large loads and optimise the
job's performance, the following strategy is used:

We search for all jobs in the interval between a week and one hour ago,
and one week ago. All finished tasks from this period are guaranteed to
be older than one week.

This search strategy allows us to minimise the amount of data fetched
from the database, while ensuring no files of old finished tasks remain
on the filesystem.