Added an alternative `update` method
The main purpose of this service method is to contain the setters within
the synchronized block. As the method is called with a Task object, a
method reference to a setter, and the new value passed to the setter,
this means that the object value is changed within the synchronized
block. This might not be necessary, but I think it can have some
applications.