Reimplemented scheduled tasks
The `TaskExecutor` and `TaskResultCleaner` have been reimplemented as
`TaskRunner` and `TaskCleaner` respectively. The new classes both
implement the `Runnable` interface, with both classes being registered
in the `SchedulerConfig` during startup. The two are now also concrete
classes, and as a result, the autowired beans they use have to be
provided by the configurer during instantiation. Another change
resulting from this is the fact that `TaskRunner` now uses explicit
transaction execution through a `PlatformTransactionManager`, as the
`@Transactional` annotation can only be used for beans. The benefit of
explicit transaction management is that we can more easily and reliably
cancel tasks. We once again rely on `version` attribute mismatches,
however since these exceptions are now localized entirely within the
read transaction of `TaskRunner`, this means that we no longer have to
use the post processor for the `@Scheduled` annotation during calls
to the cancel method.