Added `expired` column to `Task` which will be used in `TaskCleaner`
This column indicates whether the result files related to this task are
still available. By default, all tasks whose status is contained within
the ACTIVE category have this value set to false (and in these states,
it can not be true). The INACTIVE category has some additional caveats:
tasks terminating with an ERROR or CANCELLED status have their
corresponding files immediately removed, allowing us to ignore them
during the cleaning process, and set the flag immediately to true. As it
stands, these statuses can not have a false expired value. Lastly, the
FINISHED status can have both true and false values, depending on when
the task terminated. By iterating over all the non-expired tasks, the
cleaner will determine for which ones it needs to flip the expired flag,
and delete the corresponding files.