Overhauled `SchedulerErrorHandler`
The responsibility of recording task errors and cleaning up files of
failed tasks has been shifted away completely from the `TaskRunner` and
moved into the `SchedulerErrorHandler`. Task failure is now recorded
through a dedicated method in the `TaskService`, which apart from
marking a task as failed, also converts the stack trace to a string and
stores it along with the task.