Simplified the `PostConstruct` procedure of the main class
Now we simply suppress the IOException, rather than swallowing it and
suffering the consequences later. The exception in question is only
thrown if we can not create the temporary directory for the platform on
the filesystem. In such a case, the server should not be running anyway,
as all the components that rely on the `FileSystemService` are going to
experience failures.