Added service and repository layer for `Configuration`
Our application will have its own dedicated environment, residing below
the `systemEnvironment` created by Spring. The goal of this service is
both to act as an easy accessor for the environment, as well as to
modify the environment property values, and ensure that they are
persisted in the database. In its current form, it supports the
following operations:
1) Checking if a configuration property exists in the dl4se-specific
   environment layer. The database does not need to be used here, as
   we will soon define a bean that copies the DB contents into the
   environment on startup
2) Getting the value of a configuration property, all the while casting
   it to the type specified in the parameter.
3) Modifying the value of a configuration property. The configuration is
   first saved to the database, following by reloading the application-
   specific environment. In this way, we ensure that the environment
   state is persisted, and the changes are visible to classes depending
   on the configuration property values.

Although manual tests of the service reveal that it does indeed work as
intended, there are currently no concurrency settings.

The configuration repository is simplistic, providing only the basic
functionalities from JpaRepository.