Added `ReentrantReadWriteLock` to `ConfigurationService`
For our configuration service, we want changes to be immediately visible
in the application runtime. As such, we should prevent scenarios where
a value is returned while modifications to the environment are being
made. Furthermore, we should ensure that environmental changes are
limited to one at a time. In this scenario, a reentrant lock with
separate read and write locks is best suited to the job. The service
`get` method acquires a read lock, while the `modify` method acquires a
write lock. While any number of threads can acquire a read lock, as soon
as a write lock is acquired, we suspend all reads and writes until the
lock is released.