Added `TaskExecutorRecoveryBean`
Running only once on startup, this bean fetches all tasks that are in
the EXECUTING state. Tasks with such a state on startup are the result
of an unexpected shutdown. Should not break the current task-selection
setup, since it is single-threaded, but I should improve the task
selection procedure to be less complicated.