Updated synchronization for `getNext`
Due to transactions not being committed by one thread before the
invocation lock is acquired by the other thread, we ended up in
concurrency-related exceptions being thrown infrequently. By moving the
synchronization to the caller (`TaskExecutor`), we solve this problem,
as by the time the lock is released by the executor, the service method
transaction will have been committed.