Added Login Page
On the route `/login`, the user can input their information to submit a
login request. For now, the token returned by this request is just
logged out to the console. Other steps, such as clearing the form,
redirecting the user to his profile etc. will have to be done in the
future. We also introduce a dedicated component: `TextInputForm`, which
apart from the login, will also be used for the registration.