Added explicit validation for form inputs
Along with all the input specification props, we also pass validators.
Validators are boolean functions used to validate the input in the
component, which may optionally display message hints to aid validation.
While message hints are disabled for the login, we will enable the for
the registration.

This change also sees the removal of the `required` field, as the usage
of a validator can effectively replace it.