Authentication token is now preserved for page reloads.
Using the vuex-persist plugin, we store the retrieved token during login
so that it may be used in other parts of the application. Pages such as
`login`, `profile` and the task creation page should perform checks on
the value of the token by making API calls during page load. This way,
if a token has expired it can be cleared, forcing the user to log back
in once more.