Added verification view
Added the view that will make the necessary API calls to verify a user
based on the provided token information. In its current state, the view
makes these API calls on view creation. Depending on the status code:
- 200: The token exists and is valid, the verification is finished, and
       the user is redirected home, presented with a success message;
- 403: The token is expired and the page content is rendered, allowing
       the user to either request a new one or return home;
- 404: The token does not exist, so the user is redirected home, and
       presented with the appropriate error.

Note that the link sent via email still points to the API endpoint. The
next commit will change this, as it will construct the link to the
front-end page rather than the API. In this case, we may not even need
to send the resend link in case of 403, as we can just construct it on
the front-end side.