Added generic 404 page
The `NotFoundView` is used as a fallback view for when the client
requests a page whose route is not explicitly covered. It is essentially
a fallback route. It only displays the error code, the error title, and
optionally a description as well. It also contains a navigation link for
returning home.