`SectionRepo` now displays invalid feedback message for language select
The `dropdownState` computed property is used for determining whether
to display the feedback. Note that a `undefined` check needs to be
performed so that the computed property does not return errors. Logging
has revealed that the value is computed twice on startup: yielding
`undefined` the first time, and then the state of the `DropdownSelect`
child. The only explanation I have for this is: first the parent
requests the reactive vuelidate instance of the child that has not
rendered yet. Once it does render however, its reactive nature forces
the computed property to recompute, thus yielding the state value.