Added external `validators` to `Counter`
This prop is merely an object of `v$` instances passed to the `Counter`
from the parent component. When non-empty, if any single `v$` instance
is `$invalid`, then it will override the validity state of the `Counter`
forcing it to display as invalid. The `Counter` will technically not
have its `$invalid` set to true, but the parent component does. This was
the only simple way I could come up with to do it. If I find a better
way, such as passing the validators as props, then I will change how
this validation works.