Simplified calculation logic behind `anyRequired` and `entryRequired`
Both now look for a key with the name `required` in the `rules`.
This means that we can use different types of functions. This is
especially important when we start customizing the built-in functions
to display custom messages. That being said, as long as any validator
is tied to the field with the required name, then the `required`
asterisk will render.