Added default parameter settings to `CodeProcessingDto`
This makes it so that all the parameters in the aforementioned payload
object are optional, and have default values. As a result, the idioms
will for example always default to an empty array, regardless of whether
we specify them as null in the request body JSON, or don't specify the
parameter value at all. Similar reasoning applies to the `remove_X`
fields: both will default to false if null or unspecified. The `mask_X`
fields will always default to Java's default field of Object type
setting: null.