Added new constraint: `AllNullOrNotNull`
From a provided String list of field names, the constraint checks if:
1) All the field values are non-null
2) All the field values are null

If neither of these conditions are met, then a constraint violation is
thrown. This constraint will be used to check the masking parameters in
code processing, as they must all either be set to a value, or be null.