Adding `StreamSaver.js` to the dependency list
This library will be used to consume the back-end supplied stream, which
means that the front-end will not have to load the entire stream into
memory, but rather consume the stream and pipe the response contents
directly to the filesystem.

This also comes with `web-streams-polyfill`, to allow the same
dependencies to be used in Firefox/Safari.