Added `DownloadView`
A page that serves as a middleman between the client browser and the
server, when downloading task files. Uses the previously introduced
library to pipe the back-end supplied content, directly into a file on
the filesystem. This way, even files as large as several gigabytes can
be downloaded without being loaded into memory by the browser.