Refactored Task and User services and controllers to use Pageable
Building upon the previous commits, we introduce the `Pageable` object
parameter for various paged API endpoints. If paging parameters are not
specified, the default page is the first one limited to 20 results. The
sorting direction for each of these endpoints differs on a per-endpoint
basis (i.e. the default sort column for paged task retrieval differs for
querying users and tasks). The default sort direction however is always
descending, regardless of the column used for sorting.