Pagination now starts at 1 rather than 0
This is to be more in line with most other APIs out there. GitHub's API
for instance does one-indexed pagination. This change also brings our
back-end closer in-line to the front-end. Note that the back-end
internally still interprets `page=1` as page 0. It's just that data sent
from and to the front-end will be higher by 1.