Added `Abbreviation`
This component provides a way to abbreviate data, while still preserving
the original content, optionally displaying it on demand. More
specifically, it accepts the value in question, and a transformation
function for said value. The transformed value is displayed in a `span`
tag, while the original is displayed in a tooltip on hover. Note that
the transformation function can also return an HTML String to the
tooltip allowing for more versatile display capabilities.