Created centralized `axios` instance creation file
This allows us to create a single instance with our own custom settings.
The next commits will introduce the environment settings, as well as
interceptors for this new `axios` instance.

Note that in order to make our components as reusable as possible, we
should refrain from using `axios` outside views. Components should only
accept providers or consumers as props, which are passed by views, and
which define how the API calls are made in the first place.