Refactored `TextInputForm`
The component no longer has a responsibility of making POST requests
on submission. Instead, the view that uses said component must also
define a (preferably async) consumer function, and pass it down to
the component as a prop. The result of this refactoring is a more
versatile and reusable component, that is completely de-coupled from
the `axios` client.