Refactored `PaginatedTable`
Same deal as was the case with `TextInputForm`. The containing view is
now responsible for declaring the table item provider. Unfortunately,
this does come at the cost of having to supply the `total-items` as a
prop, which will be used by Bootstrap's `Pagination` component.

As a general rule of thumb, the parent's passed provider should also
have the responsibility of updating the data passed to `total-items`
inside the parent.