`HomeView` now performs a connection check before being mounted
Rather than passing a prop to determine whether a message is displayed,
we instead perform an API can during the `beforeMount` hook. While this
does not solve the problem of the message not displaying after
non-navigation, it does simplify router navigation (we don't need to
pass a prop anymore). In a subsequent commit, this hook will also set
the general disabled state for the page, allowing us to prevent site
access in case the server is down.