Added disabled state and better styling for `HomeView` links
Card links now use the `card-link` class provided by bootstrap. However,
we perform a small override, whereas in case the link is disabled, the
link text does not darken.

Furthermore, not all navigation will be disabled in case the servers are
down. I intend to make the pages that are not server-reliant visitable.