Added admin check for `DashboardView`
Before rendering the dashboard, we use the API to check if the logged
user is an admin. Rendering is delayed until the check is complete.
The outcome of this check is used to set a flag in the view's state,
which later determines which API endpoints are used, as well as which
components are rendered.