Added `RepoMaintainer`
This scheduled job will periodically stream over our list of mined
repositories. If a repository's visibility status changes, the
`is_deleted` flag is set to reflect that. This is also why we iterate
over all repositories, as opposed to just the deleted ones: certain
repos may become publicly available again, and in such case the flag
should be reset.

Yes I'm aware the `is_deleted` name is not the most reflective, but
neither is `is_visible` nor `public`. Maybe `available` would be better?