Added `DynamicUpdate` annotation to `GitRepo`
In this way, the generated SQL statements will only affect the modified
columns, as opposed to updating all of them (which is what the cached
statement does by default).

Although this may allegedly cause a performance hit, local testing has
show that the `Crawler` and `RepoMaintainer` work the same as before.
Whether this change will have any long-lasting effects remains to be
seen.