Overhauled hierarchy for processing
Fixes #12

Move over `Abstracted`, here comes `Processed`. By using this more
general wrapper class, we ensure that any type of processing is
reversible. This is done by attaching the processed code/ast version as
a supplement to the `Code` instances. Therefore, if we apply multiple
types of processing like token masking and comment removal, the clients
would be able to see the original version alongside the processed
result. Naturally, this does come at a cost of increased file size.
The benefits of this approach should not be dismissed:
1) No matter how many new processing operations we introduce, the
   abstraction remains the same;
2) We alleviate the client from performing any of the reconstructions
   themselves. Rather than reverse each processing operation, they have
   direct access to the original version of the code.