The `id` field of `Code` is now ignored at the export level
This is because we use an `ObjectMapper` to construct the `Code` entity
`Map` representation. At that point, we still need the `id` as we are
streaming over the database. The field is now only removed prior to
being written. In the future, we will migrate all the Json rule
annotations into separate configurations.