Overhauled token masking
Fixes #12

Token masking has seen a complete makeover, where apart from the masked
`content`, the clients are also provided with an `oracle`: the solution
to the masking. In order for the masking to be solvable, we had to
enumerate each of the masked tokens. This means that if the client
specifies the masking token as `extra_id`, the replacement tokens
will be `<extra_id_1>`, `<extra_id_2>`, etc. Although the resulting
masked source code will be single-line, we have configured the masker
to reconstruct whitespaces as faithfully as possible.