Bugfix: Parameter validation for `ExtensionBasedFileVisitor`
There were no guards for including `null` or empty/blank arguments for
this visitor. This commit introduces preconditions that will check the
provided arguments, and throw the appropriate exception in case they
are invalid. Because these checks need to be performed before any
actual instances are constructed, instance generation had to be
relegated to a static factory.