Added new validation constraint: `PositiveOrZeroOrNull`
We will need this constraint when we introduce the `size` column for
task. The reason we can not use just `PositiveOrZero` is because we want
to distinguish between a file with no contents (0 bytes in size) and a
file that does not exist (whose size is currently null). As it stands
there are no validation constraints that allow this, so we need our own.