Introducing `formatterMixin`
This mixin contains the single `format` method. It is a generic and
highly configurable method that allows any number to be formatted
according to a unit scale. AS it stands, we will use it to format the
created dataset sizes, as well as the ticks of the graphs within the
`StatsView`.