Splitting the old task download endpoint into two new ones
In the old API, the user would submit the access token from the store in
order to download the dataset file. Although straightforward, the
problem with this approach lay in the fact that downloads could not be
made with the browser natively, but rather through the XMLHttpRequest or
the fetch API. We do a mix of both now as follows: the download method
itself is public, but keeping it secured involves obtaining a special
token through an API endpoint that requires authorization. After the
token has been obtained, we include it in the download request as a
parameter. Using it in the parameter, as opposed to the request body
means that we can `href` to the resource directly.