Improvements to `logOut` store action
By using the `target` query parameter, we now direct from the `/login`
page to other pages dynamically. The homepage is the default option. At
the same time, this allows us to redirect a use back to a restricted
page in case of token expiry. Lastly, logging in/out on a public page
causes no page change to occur.