`BackToTop` now scrolls all the way back to the top of the page
This involves a "hacky" solution by virtue of the offset. Since the
library does not allow to scroll by selectors other than the `id`, and
since I do not want to introduce coupling by referencing elements by
`id` from separate files, I instead use the offset to scroll the desired
element into view. Note that if the need arises to introduce the button
to other views, it will be moved to `App.vue` and use logic to determine
when to render. At that point the offset hack will no longer be
required.