The `PaginatedTable` now uses a prop to dynamically generate the slots
In essence, one needs to define the slot names as an array of strings.
The table then assigns each slot under the name `controls({name})`.
Similarly to how the `BTable` component offers custom column rendering.
The reason why this was introduced in the first place was to provide
wrappers that would more cleanly separate the individual sections of
controls.