First draft of the `Analyzer` hierarchy
Meant to serve as a replacement for the existing `Parser` hierarchy from
the `dl4se-crawler` module. In the report I mentioned how it was a
misnomer, and all I have done thus far was done with the intent to
address this issue. The beauty of this new interface and subtypes is
that it will be completely modular. While a lot of operations are
generalizable (thanks to tree-sitter), all steps within the `Analyzer`
related to constructing the requisite entities will allow redefinitions
for a language-specific `Analyzer`.