The `JavaAnalyzer` information now takes previous comments into account
Limitation of initial implementation: while the comments were
concatenated the function source, the rest of the statistics were not
updated to match. With this comment the `JavaAnalyzer`, which is
currently the only analyzer requiring these actions, was updated to
perform additional recalculations for the previous sibling comments.

As a result of this change, the S-Expression field is now wrapped with
an additional wrapper node. While not a part of the tree-sitter grammar,
it is nonetheless needed to provide a valid structure that will allow
the expression to be queried in the future.