Added another layer of abstraction: `PreviousCommentsAnalyzer`
Terrible name for a class. Was never good with naming things. Don't be
surprised if this gets changed down the line. Anyway, this new abstract
class isolates function extraction logic that will be specific to
languages whose comments and documentation are defined as extras that
by convention precede the declaration. In other words the existing
`JavaAnalyzer` as well as those that will be written for C, C++ and
JavaScript. Python is (un?)fortunately a completely different can of
worms.