Added `External` supertype
The `QueryCursor`, `Tree`, `TreeCursor` and `Parser` classes now share
a common superclass called `External`. As the name suggests, this
abstract class signifies that its concrete descendants are not pure
Java classes. What this means is that each subclass is direct mapping
to a C struct. Such classes are defined by a common set of
characteristics:

- Each has a pointer that is used to directly access the struct
- This pointer is `final`, as it should never be changed
- Null instances are defined by a 0 pointer value
- All instances can be checked if they are null

`External` itself is defined as package private, as this is a rare case
where the substitution principle is not desired. This may be subject to
change in the future, provided that different needs arise. For instance,
we may keep references of all `External` instances in an ordered
collection and then free the memory associated with then en-masse.