HUGE COMMIT, READ MESSAGE BODY!
I have begun introducing the `PythonAnalyzer` into the `Analyzer`
hierarchy. This implies a number of changes:
- `Validated` has been completely removed
- `Queries` now offers only a `validate` and abstract `execute` methods
- Implementations have been divided into single- and multi-capture
- `PreviousCommentsAnalyzer` has been completely removed

Q: What is the difference between single- and multi-capture `Queries`?
A: Single-capture `Queries` will be used to locate independent source
   nodes like comments, while multi-capture `Queries` will be used to
   locate nodes whose immediate neighbors will also be stored along
   with the node itself. In the case of Java, comments preceding
   callable declarations will be extracted alongside the declaration,
   while in the case of Python, all decorators preceding the function
   definition will also be extracted.

Q: Why was `Validated` removed?
A: Thanks to the introduction of "count" methods to the `Query` class,
   as well as the inclusion exceptions during construction, I believe
   that the interface is no longer needed, and the logic can be
   extracted to the common ancestor of the implementing classes, which
   in this case is `Queries`.

Q: Does this also mean that we will remove `Traverser`?
A: No, it may be useful for the `PythonBoilerplateEnumerator`, which I
   have yet to implement. Doing so requires first a discussion on what
   constitutes "boilerplate" in Python.

Q: Why was `PreviousCommentsAnalyzer` removed?
A: Thanks to the introduction of multi-capture queries, the middleman is
   no longer required. This query type will now always be used for the
   extraction of callable declarations from files, so the logic it
   introduced has been hoisted up into the parent.