`SExpressionPrinter` adds a bracket prefix and suffix implicitly
The goal of this printing should not be to produce a valid s-expression
tree, but rather to give a valid s-expression that can be later used to
construct a `Query`. Sibling nodes can be grouped in a bracket, meaning
that we do not need to add any fake anchor or whatever. We did however
need to introduce a custom joiner that would not introduce any
parentheses in the event of empty input collection or just a single
`Node`.