Introducing `FunctionSyntaxTreePrinter` for recalculating function ASTs
The protected methods of `AbstractAnalyzer` have been moved to their own
private inner class. Why is it an inner class? To make use of the
existing members more efficiently. Why is its factory part of the
`AbstractAnalyzer` and not the class itself? Language limitations!
This will change once ported to Java 17.