Languages are now initialized by a startup task in the crawler
Rather than relying on migrations, it is tenfold better to have a
dedicated configuration file that is read each time the server starts.
In this way, we can add languages on the fly, as well as change the
supported extensions of existing languages.

This does however mean that there is a miss-match possibility between
the DB and this file, whereas entries removed from this file remain in
the DB. We should configure the crawler to only mine languages mentioned
in this file.