Added edge case to `JavaLineCounterTest`
What if the acquired nodes are on the same line? For instance: a block
comment preceding a method declaration, without any new lines in
between. In this case, we can not count their line span individually
and then add them together. What we must do instead is merge ranges
whose end position row is equal to the start position row of its
neighbor.