Bumping minor version of `java-tree-sitter`
`1.4.2` -> `1.5.1`

Since this is the version that allows direct access to content from
nodes, a lot of the analysis stuff was simplified. As of this commit,
the `NodeMapper` class has been rendered completely irrelevant. This has
led to massive improvements in the overall construction of all the
analyzer classes, as well as their inheritance hierarchy.