Refactoring `EntityNotFoundException`
The logic for constructing individual exception messages has been
extracted to the abstract class. Furthermore, all implementing instances
are associated with a specific entity type. This allows exceptions to
only be constructed with `Attribute` instances associated with the same
class. Furthermore, the value for which the entity was not found must
coincide with the value of the attribute. As a result there can no
longer be instances of type `TaskNotFoundException` that refer to a
`Language` column with a completely different type value. Things are
much more restricted now.