`Configuration` initialization is now handled by an `InitializingBean`
This is similar to how languages are initialized in the crawler.
It also makes more sense to do it this way, since the migrations
should not be used to insert data into tables. That being said,
the configurations will only configure the number of running
tasks and maximum number of user requests for the time being.
I believe that the CRON expressions are not helpful enough to
be configurable.