Bugfix: incorrect checkpoint restart in case of sudden downtime
We only mutate the converted specification to resume the scroll from the
specified checkpoint ID, only after we have finished counting the number
of entities that will be exported. The reason we cannot do this in the
converter is because setting the checkpoint there would cause the count
to shrink.