Fix: Include `File` path in `Function` exports
Equivalent to a43aa1213c8211143632f0cd8d97ca5b05ae363a on main branch.
This is the "Jackson" way of doing it as opposed to using a transient
getter. I'm not sure if this is entirely better because it relies on the
schema not being changed constantly. But at the same time, the other
approach has the drawback of having to define a getter for every element
we wish to expose. Here, we just have to add or remove elements to this
array if we wish to ignore or un-ignore them respectively.